Add AcceptItem processing to NCIP::ILS::Evergreen.
This necessitated the addition of several methods to
NCIP::ILS::Evergreen for looking up users, copies, etc., as well
as for the placing of holds, and checking for some errors. Most
of the new methods will be useful for other message handlers.
The LookupUser handler was also modified to use the new methods
where appropriate.

AcceptItemResonse.inc was also modified to take into account the
new way of passing data to templates. NCIP::ILS::Koha has yet to
be modified to use these new methods.

Signed-off-by: Jason Stephenson <jason@sigio.com>