Add $code option to NCIP::ILS->find_bibliographic_id.
I wonder now, if I should not have overriden the method in the
Evergreen driver, but I thought this would be generically useful.

We will probably want a method that can return all of the IDs at
some point, but for now, any subclasses that need that, can
implement it, themselves.

Signed-off-by: Jason Stephenson <jason@sigio.com>