Templates changes for Auto-graphics.
They requested that we remove "carriage returns" from the output. I'm not
100% certain if he meant all carriage returns or if he just wanted the
weird spacing that some elements had where the open tag is on one line
followed by the value and the close tag on the next line. I don't really
know exactly why that was happening, but it seemed to have something to
do with trying to make the value come out on a line by its own between
the open and close tags. This change puts those elements on one line.

Signed-off-by: Jason Stephenson <jason@sigio.com>