Modify NCIP::ILS::Evergreen based on communication and testing.
Auto-Graphics is sending our target branch code in the ToAgencyId
field of the InitiationHeader.  They are not sending some of the
optional fields that we are looking for.  Most of the changes in
this commit revolve around using the ToAgencyId as a fallback
when the other fields are not available.  The focus of these
changes are the AcceptItem and RequestItem messages since the
local location is most important there.

Signed-off-by: Jason Stephenson <jason@sigio.com>