Reimplement requestitem handler in NCIP::ILS::Evergreen.
Testing with Auto-Graphics suggests that our original logic was too
complicated, so we simplify it to only handle what Auto-Graphics
sends us.  We may want to implement a more generic handler in the
future when NCIPServer is used with more vendors.  However, it seems
fairly typical for responders to also be vendor specific.

While simplifying the logic, this commit gets rid of some unneeded
funtions.

We also fix a typo in the RequestItemResponse.inc template.

Signed-off-by: Jason Stephenson <jason@sigio.com>